requirements.txt
Flask
requests
beautifulsoup4
selenium
webdriver-manager
Pillow
colorgram.py
reportlab
cairosvg
cssutils
google-generativeai